code fix and optimzation

+++ src/main/java/com/learning/product/service/ProductService.java
@@ -10,6 +10,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class ProductService {
@@ -19,38 +20,44 @@ public class ProductService {
     public ProductService(ProductRepo productRepo) {
         this.productRepo = productRepo;
     }
-    public ProductResponse addProduct(ProductRequest productResponse) {
+    @Transactional
+    public ProductResponse addProduct(ProductRequest request) {
         log.info("addProduct");
-        Products product = new Products(productResponse.getProductName(),productResponse.getProductDescription(),
-                productResponse.getProductPrice(),productResponse.getProductQuantity());
+        Products product = new Products(request.getProductName(),request.getProductDescription(),
+                request.getProductPrice(),request.getProductQuantity());
         return mapToResponse(productRepo.save(product));
     }
-    public ProductResponse updateProduct(Long id, ProductRequest productResponse) {
+    @Transactional
+    public ProductResponse updateProduct(Long id, ProductRequest request) {
        log.info("updateProduct");
        Products product = productRepo.findById(id).
                orElseThrow(()-> new ProductNotFound("Product not found"));
-        product.setName(productResponse.getProductName());
-        product.setDescription(productResponse.getProductDescription());
-        product.setPrice(productResponse.getProductPrice());
-        product.setQuantity(productResponse.getProductQuantity());
+        product.setName(request.getProductName());
+        product.setDescription(request.getProductDescription());
+        product.setPrice(request.getProductPrice());
+        product.setQuantity(request.getProductQuantity());
         return mapToResponse(productRepo.save(product));
     }
-
+    @Transactional
     public Long deleteProduct(Long id) {
+        if (!productRepo.findById(id).isPresent()) {
+            throw new ProductNotFound(
+                    "Product not found with id: " + id);
+        }
         log.info("deleteProduct");
         productRepo.deleteById(id);
 
         return id;
 
     }
-
+    @Transactional(readOnly= true)
     public ProductResponse getProductById(Long id) {
         log.info("getProductById");
         Products product= productRepo.findById(id).
                 orElseThrow(()-> new ProductNotFound("Product not found"));
         return mapToResponse(product);
     }
-
+    @Transactional(readOnly= true)
     public Page<ProductResponse> getAllProducts(Pageable pageable) {
         log.info("Fetching products. Page: {}, Size: {}",
                 pageable.getPageNumber(),
@@ -67,6 +74,25 @@ public class ProductService {
                 product.getQuantity()
         );
 
+    }
+    public Page<ProductResponse> searchProducts(
+            String name,
+            Pageable pageable) {
+
+        log.info(
+                "Searching products with name: {}, page: {}, size: {}",
+                name,
+                pageable.getPageNumber(),
+                pageable.getPageSize()
+        );
+
+        Page<Products> productsPage =
+                productRepo.findByNameStartingWithIgnoreCase(
+                        name,
+                        pageable
+                );
+
+        return productsPage.map(this::mapToResponse);
     }
 }
